+++ CHANGELOG.md
@@ -1,10 +1,20 @@
+## 0.9.60 Release:
+## New Content:
+- Halfling: Brave and Lucky
+- Mark of Healing Halfling: Medical Intuition
+- Mark of Hospitality Halfling: Ever Hospitable
+- Stout Halfling: Stout Resilience
+## Bug Fixes:
+- Create Pact Weapon
+- Stone's Endurance
+## Update Notes:
+- Support for D&D system version 3.0.0 will be in a future update. Do not update your system yet.
+  
 # 0.9.59 Release:
 ## Bug Fixes:
 - Document Selector Dialog
 - Rage
 - Enlarge/Reduce
-## Update Note:
-- Support for D&D system version 3.0.0 will be in a future update. Do not update your system yet.
   
 # 0.9.58 Relaese:
 ## Bug Fixes: